+++ jgclark.DashboardReact/src/react/components/Header.jsx
@@ -85,7 +85,7 @@ const Header = ({ lastFullRefresh }: Props): React$Node => {
       <RefreshControl refreshing={pluginData.refreshing === true} handleRefreshClick={handleRefreshClick} />
       
       <div className="totalCounts">
-        <span id="totalDoneCount">0</span> items closed
+        {/* <span id="totalDoneCount">0</span> items closed */}
       </div>
       <DropdownMenu items={dropdownItems} handleSwitchChange={handleSwitchChange} className={'filter'} />
       {/* TODO(later): more detailed setting dialog, using className={'settings'} and icon fa-gear */}